Refactored what I wrote to be a little bit more expandable.
